modified all tests and pom classes
