+++ src/test/java/ru/praktikum_services/qa_scooter/electivetests/ReferencesTest.java
@@ -1,10 +1,12 @@
-package ru.praktikum_services.qa_scooter.electiveTest;
+package ru.praktikum_services.qa_scooter.electivetests;
 
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import ru.praktikum_services.qa_scooter.pom.OrderPage;
 
 import static org.junit.Assert.assertEquals;
 
@@ -12,12 +14,17 @@ public class ReferencesTest {
 
     private WebDriver driver;
 
-    @Test
-    public void samokatLogoRef() {
+    @Before
+    public void start() {
         ChromeOptions options = new ChromeOptions();
         driver = new ChromeDriver(options);
-        driver.get("https://qa-scooter.praktikum-services.ru/order/");
+    }
+
+    @Test
+    public void samokatLogoRef() {
+
         OrderPage objOrderPage = new OrderPage(driver);
+        objOrderPage.getUrl();
         objOrderPage.samokatLogoClick();
 
         assertEquals("Перехода на главную страницу Самоката по клику на логотип Самокат нет", "https://qa-scooter.praktikum-services.ru/", driver.getCurrentUrl());
@@ -25,12 +32,11 @@ public class ReferencesTest {
 
     @Test
     public void yandexLogoRef() {
-        ChromeOptions options = new ChromeOptions();
-        driver = new ChromeDriver(options);
-        driver.get("https://qa-scooter.praktikum-services.ru/");
+
         OrderPage objOrderPage = new OrderPage(driver);
+        objOrderPage.getUrl();
         objOrderPage.yandexLogoClick();
-        objOrderPage.waitForLoadPage();
+        objOrderPage.waitForLoadYandexPage();
 
         assertEquals("Перехода на главную страницу Яндекса по клику на логотип Яндекса нет", "https://yandex.ru/", driver.getCurrentUrl());
     }
