+++ src/test/java/ru/praktikum_services/qa_scooter/electivetests/WrongNumberPageTest.java
@@ -1,10 +1,12 @@
-package ru.praktikum_services.qa_scooter.electiveTest;
+package ru.praktikum_services.qa_scooter.electivetests;
 
 import org.junit.After;
 import org.junit.Test;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import ru.praktikum_services.qa_scooter.pom.MainPage;
+import ru.praktikum_services.qa_scooter.pom.StatusPage;
 
 import static org.junit.Assert.assertEquals;
 
@@ -13,13 +15,14 @@ public class WrongNumberPageTest {
     private WebDriver driver;
 
     @Test
-        public void CheckWrongNumberStatus() {
+        public void checkWrongNumberStatus() {
 
         final String orderNumber = "4";
         ChromeOptions options = new ChromeOptions();
         driver = new ChromeDriver(options);
-        driver.get("https://qa-scooter.praktikum-services.ru/");
+
         MainPage objMainPage = new MainPage(driver);
+        objMainPage.getUrl();
         objMainPage.orderStatusButtonClick();
         objMainPage.waitForOrderNumberField();
         objMainPage.setOrderNumber(orderNumber);
@@ -29,7 +32,6 @@ public class WrongNumberPageTest {
         boolean actual = objStatusPage.checkNotFoundPic();
 
         assertEquals("Нет надписи, что заказа нет", true, actual);
-
     }
 
     @After
